+++ src/Controller/OrderController.php
@@ -6,6 +6,7 @@ use App\Entity\Product;
 use App\Repository\OrdersRepository;
 Use App\Repository\ProductRepository;
 use App\Repository\CustomerRepository;
+use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -37,28 +38,34 @@ class OrderController extends AbstractController
      */
     public function addOrder(Request $request) : JsonResponse
     {
-        $data = json_decode($request->getContent(), true);
-        // $data = json_decode($request->getContent(), true);
+        try {
+            $data = json_decode($request->getContent(), true);
+            // $data = json_decode($request->getContent(), true);
+    
+            // $temp = $request->getContent();
+            // $temp = json_decode($temp);
+        
+            $customer_id = $data["customer_id"];
+            $product_id = $data["product_id"];
+            $price = $data["price"];
+            $customer = $this->customerRepository->findOneBy(['id'=>$customer_id]);
+            $product = $this->productRepository->findOneBy(['id'=>$product_id]);
+            // $product = $this->productRepository->findOneBy(['id' => $product_id]);
+            
+            
+            // $price = $product->getPrice();
+            
+            if (empty($customer_id) || empty($product_id) || empty($price) ){
+                throw new NotFoundHttpException("Expecting mandatory parameters:");
+            }
+            
+            
+            $id = $this->orderRepository->saveOrder($customer, $product);
+            return new JsonResponse(['message' => "order created with id number ".$id], Response::HTTP_CREATED);
 
-        // $temp = $request->getContent();
-        // $temp = json_decode($temp);
-        $customer_id = $data["customer_id"];
-        $product_id = $data["product_id"];
-        $price = $data["price"];
-        $customer = $this->customerRepository->findOneBy(['id'=>$customer_id]);
-        $product = $this->productRepository->findOneBy(['id'=>$product_id]);
-        // $product = $this->productRepository->findOneBy(['id' => $product_id]);
-        
-        
-        // $price = $product->getPrice();
-        
-        if (empty($customer_id) || empty($product_id) || empty($price) ){
-            throw new NotFoundHttpException("Expecting mandatory parameters:");
+        } catch(Exception $e){
+            return new JsonResponse(["message" => $e->getMessage()]);
         }
-        
-        
-        $id = $this->orderRepository->saveOrder($customer, $product);
-        return new JsonResponse(['message' => "order created with id number ".$id], Response::HTTP_CREATED);
     }
 
     /**
@@ -66,18 +73,28 @@ class OrderController extends AbstractController
      */
     public function getOneOrder($id): JsonResponse
     {
-        $order = $this->orderRepository->findOneBy(['id' => $id]);
-        $customer_name = $order->getCustomerId()->getFirstName();
-        $product_name = $order->getProductId()->getName();
-        $data = [
-            'id' => $order->getId(),
-            'customer' => $customer_name,
-            'product' => $product_name,
-            'price' => $order->getProductId()->getPrice(),
-            
-        ];
+        try {
 
-        return new JsonResponse(['order' => $data], Response::HTTP_OK);
+            $order = $this->orderRepository->findOneBy(['id' => $id]);
+
+            if ($order == null){
+                return new JsonResponse(["message" => "Order does not exist"]);
+            }
+
+            $customer_name = $order->getCustomerId()->getFirstName();
+            $product_name = $order->getProductId()->getName();
+            $data = [
+                'id' => $order->getId(),
+                'customer' => $customer_name,
+                'product' => $product_name,
+                'price' => $order->getProductId()->getPrice(),
+                
+            ];
+    
+            return new JsonResponse(['order' => $data], Response::HTTP_OK);
+        } catch(Exception $e){
+            return new JsonResponse(["message" => $e->getMessage()]);
+        }
     }
 
     /**
@@ -85,12 +102,21 @@ class OrderController extends AbstractController
      */
     public function deleteOrder($id) : JsonResponse
     {
-        $order = $this->orderRepository->findOneBy(['id'=>$id]);
-        $this->orderRepository->removeOrder($order);
+        try {
 
-        return new JsonResponse(["message"=>"order deleted with id number ".$id]);
+            $order = $this->orderRepository->findOneBy(['id'=>$id]);
+    
+            if ($order == null) {
+                return new JsonResponse(["message" => "Order does not exist."]);
+            }
+    
+            $this->orderRepository->removeOrder($order);
+    
+            return new JsonResponse(["message"=>"order deleted with id number ".$id]);
 
-
+        } catch(Exception $e) {
+            return new JsonResponse([json_decode($e->getMessage())]);
+        }
     }
 
     /**
@@ -98,28 +124,36 @@ class OrderController extends AbstractController
      */
     public function updateOrder($id, Request $request) :JsonResponse
     {
-        $order = $this->orderRepository->findOneBy(["id"=>$id]);
-        // $data = json_decode($request->getContent(), true);
-        $data = json_decode($request->getContent(), true);
+        try {
+            $order = $this->orderRepository->findOneBy(["id"=>$id]);
+            // $data = json_decode($request->getContent(), true);
+    
+            if ($order == null) {
+                return new JsonResponse(["message" => "Order not found"]);
+            }
+    
+            $data = json_decode($request->getContent(), true);
+    
+            if (!empty($data['customer_id'])){
+                // Response
+                $data["customer_id"]=$this->customerRepository->findOneBy(["id"=>$data["customer_id"]]);
+            }
+            if (!empty($data['product_id']))
+            {
+                $data["product_id"] = $this->productRepository->findOneBy(["id"=>$data["product_id"]]);
+            }
+    
+            if (!empty($data['price']) )
+            {
+                $data["price"] = $this->productRepository->findOneBy(["id"=>$data["product_id"]]);
+            }
+    
+            $this->orderRepository->updateOrder($order, $data);
 
+            return new JsonResponse(["message"=> "Order updated with id number ".$id]);
 
-
-        if (!empty($data['customer_id'])){
-            // Response
-            $data["customer_id"]=$this->customerRepository->findOneBy(["id"=>$data["customer_id"]]);
+        } catch (Exception $e) {
+            return new JsonResponse(["message" => $e->getMessage()]); 
         }
-        if (!empty($data['product_id']))
-        {
-            $data["product_id"] = $this->productRepository->findOneBy(["id"=>$data["product_id"]]);
-        }
-
-        if (!empty($data['price']) )
-        {
-            $data["price"] = $this->productRepository->findOneBy(["id"=>$data["product_id"]]);
-        }
-
-        
-        $this->orderRepository->updateOrder($order, $data);
-        return new JsonResponse(["message"=> "Order updated with id number ".$id]);
     }
-}
+ }
